Refactor math
